A bit more refactoring - of the pyschemas, and made it possible test the services
