+++ common/extractors/BaseExtractor.py
