+++ common/extractors/__init__.py
@@ -1,0 +1,3 @@
+from common.extractors.LLMEntityRelationshipExtractor import (
+    LLMEntityRelationshipExtractor,
+)
