+++ copilot/app/configs
@@ -1,0 +1,1 @@
+../../configs
