--- copilot/app/supportai/extractors/__init__.py
@@ -1,3 +1,0 @@
-from app.supportai.extractors.LLMEntityRelationshipExtractor import (
-    LLMEntityRelationshipExtractor,
-)
