+++ copilot/tests/common
@@ -1,0 +1,1 @@
+../../common
