+++ docker-compose.yml
@@ -17,6 +17,20 @@ services:
     volumes:
       - ./configs/:/code/configs
 
+  eventual-consistency-service:
+    container_name: eventual-consistency-service
+    build: eventual-consistency-service
+    ports:
+      - 8001:8001
+    environment:
+      LLM_CONFIG: "/code/configs/llm_config.json"
+      DB_CONFIG: "/code/configs/db_config.json"
+      MILVUS_CONFIG: "/code/configs/milvus_config.json"
+      LOGLEVEL: "INFO"
+      MILVUS_HOST: "milvus-standalone"
+    volumes:
+      - ./configs/:/code/configs
+
   etcd:
       container_name: milvus-etcd
       image: quay.io/coreos/etcd:v3.5.5
